+++ dataset/src/blocks/DenseArray.h
@@ -40,7 +40,7 @@ class DenseArrayBlock final : public Block {
                                              /* dim= */ 1);
   }
 
-  uint32_t featureDim() const final { return _dim; };
+  uint64_t featureDim() const final { return _dim; };
 
   bool isDense() const final { return true; };
