+++ dataset/tests/MockBlock.h
@@ -15,7 +15,7 @@ class MockBlock final : public Block {
   explicit MockBlock(ColumnIdentifier column, bool dense)
       : _column(std::move(column)), _dense(dense) {}
 
-  uint32_t featureDim() const override { return 1; }
+  uint64_t featureDim() const override { return 1; }
 
   bool isDense() const override { return _dense; }
